backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
groq==0.11.0
google-generativeai==0.8.3
sentence-transformers==3.0.1
qdrant-client==1.11.0
asyncpg==0.30.0
langchain==0.3.1
langchain-community==0.3.1
langchain-text-splitters==0.3.1
pydantic==2.9.0
pydantic-settings==2.4.0
python-multipart==0.0.9